requirements.txt
appdirs==1.4.3
arrow==0.10.0
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.10.5
djangorestframework==3.5.4
nltk==3.2.2
numpy==1.12.0
oauthlib==2.0.2
packaging==16.8
pandas==0.19.2
PyJWT==1.5.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-decouple==3.0
python3-openid==3.1.0
pytz==2016.1
requests==2.13.0
requests-oauthlib==0.8.0
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
sklearn==0.0
stravalib==0.6.6
Unipath==1.1
units==0.7
uWSGI==2.0.14
